src/collecte_donnees/requirements.txt
psycopg2-binary>=2.9.1
requests>=2.26.0
msgpack-python>=0.5.6
flask>=2.0.2
